Move helpers to common location
